% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(data)
}
\arguments{
\item{data}{object}
}
\value{
data.frame
}
\description{
transforms and filters data from JSON structure to table structure
}
