\name{marsweather-package}
\alias{marsweather-package}
\alias{marsweather}
\docType{package}
\title{
\packageTitle{marsweather}
}
\description{
\packageDescription{marsweather}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{marsweather}
\packageIndices{marsweather}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{marsweather}

Maintainer: \packageMaintainer{marsweather}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
